requirements.txt
requests
PyYAML==5.1.2
distro
setuptools
salt
salt-ssh